+++ workflows/rnaseq/CHANGELOG.md
@@ -4,4 +4,8 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
  
-## [Unreleased]
+## 2025 January
+
+### Changed
+
+- `rnaseq-core` now takes an `Array[String]` for the `read_groups` param instead of a single `String` [#205](https://github.com/stjudecloud/workflows/pull/205).